@@ -18,7 +18,6 @@ classifiers =
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.5
     Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
@@ -32,7 +31,7 @@ classifiers =
 packages = find:
 package_dir =
      = src
-python_requires = >=3.5,<4
+python_requires = >=3.6,<4
 
 [options.packages.find]
 where = src
