@@ -2,102 +2,107 @@ import pytest
 
 
 # pylint: disable=attribute-defined-outside-init
-class TestOutput:
+class TestOutputBuffer:
     @pytest.fixture(autouse=True)
     def init(self, ssh_audit):
-        self.Output = ssh_audit.Output
         self.OutputBuffer = ssh_audit.OutputBuffer
 
-    def test_output_buffer_no_lines(self, output_spy):
+    def test_outputbuffer_no_lines(self, output_spy):
         output_spy.begin()
-        with self.OutputBuffer() as obuf:
-            pass
-        assert output_spy.flush() == []
+        obuf = self.OutputBuffer()
+        obuf.write()
+        assert output_spy.flush() == ['']
         output_spy.begin()
-        with self.OutputBuffer() as obuf:
-            pass
-        obuf.flush()
-        assert output_spy.flush() == []
 
-    def test_output_buffer_no_flush(self, output_spy):
-        output_spy.begin()
-        with self.OutputBuffer():
-            print('abc')
-        assert output_spy.flush() == []
-
-    def test_output_buffer_flush(self, output_spy):
-        output_spy.begin()
-        with self.OutputBuffer() as obuf:
-            print('abc')
-            print()
-            print('def')
-        obuf.flush()
-        assert output_spy.flush() == ['abc', '', 'def']
-
-    def test_output_defaults(self):
-        out = self.Output()
+    def test_outputbuffer_defaults(self):
+        obuf = self.OutputBuffer()
         # default: on
-        assert out.batch is False
-        assert out.use_colors is True
-        assert out.level == 'info'
+        assert obuf.batch is False
+        assert obuf.use_colors is True
+        assert obuf.level == 'info'
 
-    def test_output_colors(self, output_spy):
-        out = self.Output()
-        # test without colors
+    def test_outputbuffer_colors(self, output_spy):
+        out = self.OutputBuffer()
+
+        # Test without colors.
         out.use_colors = False
+
         output_spy.begin()
         out.info('info color')
+        out.write()
         assert output_spy.flush() == ['info color']
+
         output_spy.begin()
         out.head('head color')
+        out.write()
         assert output_spy.flush() == ['head color']
+
         output_spy.begin()
         out.good('good color')
+        out.write()
         assert output_spy.flush() == ['good color']
+
         output_spy.begin()
         out.warn('warn color')
+        out.write()
         assert output_spy.flush() == ['warn color']
+
         output_spy.begin()
         out.fail('fail color')
+        out.write()
         assert output_spy.flush() == ['fail color']
+
+        # If colors aren't supported by this system, skip the color tests.
         if not out.colors_supported:
             return
-        # test with colors
+
+        # Test with colors.
         out.use_colors = True
+
         output_spy.begin()
         out.info('info color')
+        out.write()
         assert output_spy.flush() == ['info color']
+
         output_spy.begin()
         out.head('head color')
-        assert output_spy.flush() == ['\x1b[0;36mhead color\x1b[0m']
+        out.write()
+        assert output_spy.flush() in [['\x1b[0;36mhead color\x1b[0m'], ['\x1b[0;96mhead color\x1b[0m']]
+
         output_spy.begin()
         out.good('good color')
-        assert output_spy.flush() == ['\x1b[0;32mgood color\x1b[0m']
+        out.write()
+        assert output_spy.flush() in [['\x1b[0;32mgood color\x1b[0m'], ['\x1b[0;92mgood color\x1b[0m']]
+
         output_spy.begin()
         out.warn('warn color')
-        assert output_spy.flush() == ['\x1b[0;33mwarn color\x1b[0m']
+        out.write()
+        assert output_spy.flush() in [['\x1b[0;33mwarn color\x1b[0m'], ['\x1b[0;93mwarn color\x1b[0m']]
+
         output_spy.begin()
         out.fail('fail color')
-        assert output_spy.flush() == ['\x1b[0;31mfail color\x1b[0m']
+        out.write()
+        assert output_spy.flush() in [['\x1b[0;31mfail color\x1b[0m'], ['\x1b[0;91mfail color\x1b[0m']]
 
-    def test_output_sep(self, output_spy):
-        out = self.Output()
+    def test_outputbuffer_sep(self, output_spy):
+        out = self.OutputBuffer()
         output_spy.begin()
         out.sep()
         out.sep()
         out.sep()
+        out.write()
         assert output_spy.flush() == ['', '', '']
 
-    def test_output_levels(self):
-        out = self.Output()
+    def test_outputbuffer_levels(self):
+        out = self.OutputBuffer()
         assert out.get_level('info') == 0
         assert out.get_level('good') == 0
         assert out.get_level('warn') == 1
         assert out.get_level('fail') == 2
         assert out.get_level('unknown') > 2
 
-    def test_output_level_property(self):
-        out = self.Output()
+    def test_outputbuffer_level_property(self):
+        out = self.OutputBuffer()
         out.level = 'info'
         assert out.level == 'info'
         out.level = 'good'
@@ -109,8 +114,8 @@ class TestOutput:
         out.level = 'invalid level'
         assert out.level == 'unknown'
 
-    def test_output_level(self, output_spy):
-        out = self.Output()
+    def test_outputbuffer_level(self, output_spy):
+        out = self.OutputBuffer()
         # visible: all
         out.level = 'info'
         output_spy.begin()
@@ -119,6 +124,7 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 5
         # visible: head, warn, fail
         out.level = 'warn'
@@ -128,6 +134,7 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 3
         # visible: head, fail
         out.level = 'fail'
@@ -137,6 +144,7 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 2
         # visible: head
         out.level = 'invalid level'
@@ -146,10 +154,11 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 1
 
-    def test_output_batch(self, output_spy):
-        out = self.Output()
+    def test_outputbuffer_batch(self, output_spy):
+        out = self.OutputBuffer()
         # visible: all
         output_spy.begin()
         out.level = 'info'
@@ -159,6 +168,7 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 5
         # visible: all except head
         output_spy.begin()
@@ -169,4 +179,5 @@ class TestOutput:
         out.good('good color')
         out.warn('warn color')
         out.fail('fail color')
+        out.write()
         assert len(output_spy.flush()) == 4
